% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{samplesPlot}
\alias{samplesPlot}
\title{Plot MCMC traceplots and density plots}
\usage{
samplesPlot(samples, var = colnames(samples), ind = NULL,
  burnin = NULL, scale = FALSE, width = 7, height = 4,
  legend = TRUE, legend.location = "topright", traceplot = TRUE,
  densityplot = TRUE, file = NULL)
}
\arguments{
\item{samples}{Array of MCMC samples, or a list of samples from multiple chains in which case the first chain is used}

\item{var}{Parameter names to plot}

\item{ind}{Chain indices to plot}

\item{burnin}{Nunber of initial MCMC samples to discard}

\item{scale}{Logical, whether to normalize each posterior chain}

\item{width}{Width of the plot}

\item{height}{Height of the plot}

\item{legend}{Logical, whether to include a legend of parameter names}

\item{legend.location}{Location of legend}

\item{traceplot}{Logical, whether to include traceplots}

\item{densityplot}{Logaical, whether to include density plots}

\item{file}{Optional filename to save figure as a file}
}
\description{
Plot MCMC traceplots and density plots
}
\examples{
samples <- cbind(rnorm(1000), rgamma(1000, 1))
colnames(samples) <- c('alpha', 'beta')
samplesPlot(samples)

}
